Generate a character vector of parameters for package options

This function generates a character vector of parameters for options defined using the custom options functions. The resulting vector can be used in documentation or for other purposes.

Usage

as_params_pkg(envir = NULL, include = NULL, exclude = NULL)

Arguments

envir

The environment containing the options. Can be NULL (default, uses the calling environment), a string (package name), or an environment object.

include

A character vector of option names to include. If NULL (default), all options are included.

exclude

A character vector of option names to exclude. Default is NULL.

Value

A character vector of parameter descriptions, where each element is a string describing an option.

Examples

if (FALSE) { # \dontrun{
# Assuming some options are defined in the current environment
params <- as_params_pkg()
cat(params, sep = "\n")

# For a specific package
params <- as_params_pkg("mypackage")
cat(params, sep = "\n")

# Include only specific options
params <- as_params_pkg(include = c("option1", "option2"))
cat(params, sep = "\n")

# Exclude specific options
params <- as_params_pkg(exclude = c("option3", "option4"))
cat(params, sep = "\n")
} # }