Check if an environment variable or option is set to a truthy value

This function returns another function that checks if a given environment variable or option is set to a value that can be interpreted as true. It first checks for an environment variable of the form R_OPTION_NAME or the exact name provided, then falls back to the value set in the specified environment.

Usage

envvar_is_true_pkg(envir = NULL)

Arguments

envir

The environment in which to look for the option. Can be NULL (default, uses .GlobalEnv), a string (package name), or an environment.

Value

A function that takes a parameter 'x' (the name of the option or environment variable to check) and returns a logical value: TRUE if the option or environment variable is set to a truthy value, FALSE otherwise.

Examples

is_true <- envvar_is_true_pkg()
Sys.setenv(R_MY_OPTION = "true")
is_true("my_option")  # Returns TRUE
#> [1] TRUE

Sys.setenv(R_MY_OPTION = "false")
is_true("my_option")  # Returns FALSE
#> [1] FALSE

Sys.setenv(MY_ENV_VAR = "yes")
is_true("MY_ENV_VAR")  # Returns TRUE
#> [1] TRUE